NetPulse/requirements.txt
colorama
speedtest-cli
requests
pyperclip==1.8.2